1/pyproject.toml
[tool.poetry]
name = "1"
version = "0.1.0"
description = ""
authors = ["SerHappy <lokhmanovkirill@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-decouple = "^3.8"

[tool.poetry.group.dev.dependencies]
mypy = "^1.7.1"
flake8 = "^6.1.0"
black = "^23.11.0"
isort = "^5.12.0"
types-requests = "^2.31.0.10"

[tool.isort]
profile = "black"
multi_line_output = 3
force_alphabetical_sort = true
force_single_line = true
lines_after_imports = 2
line_length = 120

[tool.black]
line-length = 120
target_version = ["py311"]
include = '\.pyi?$'
preview = true
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.venv
  | \.env
  | logs
)/
'''

[tool.mypy]
python_version = "3.11"
files = "*.py"
ignore_missing_imports = true
warn_return_any = false
warn_unused_configs = true
warn_unused_ignores = true
exclude = []
disallow_untyped_calls = false
disallow_untyped_defs = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
